FEAT: Add testcase & Update xocc to version 1.2.1

+++ test/exec/switch.string.c
@@ -1,4 +1,4 @@
-int ch;
+long long ch;
 int main()
 {
     ch = 'a';
